Crate problem038

Problem 38 (Pandigital multiples)

Solution detail

Since the final answer will have 9 digits, the base number that the product is formed from can have at most 4 digits. This is not many numbers to check, so for each one, form the smallest concatenated product that has at least 9 digits, then check to see if it is pandigital or not. Then just remember the largest solution that is found during this process.

Functions

answer
Solve the problem, returning the answer as a String

concatenate
Concatenate the two given numbers, a followed by b.

concatenated_product
Find the first concatenated product of n which contains at least 9 digits.

is_pandigital
Check whether the given number is pandigital.

main

num_digits
Find the number of digits in the given number.

solve
Find the largest pandigital concatenated product.